extract shared inference metrics into core microcrate
