+++ crates/bitnet-inference-metrics-core/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "bitnet-inference-metrics-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "Shared inference timing and throughput metric models"
+include = [
+  "src/**",
+  "Cargo.toml",
+]
+
+[dependencies]
+serde = { workspace = true, features = ["derive"] }
